cilium: move default route handling into route pkg
